requirements.txt
httpx==0.23.0
pycryptodome==3.15.0
